Feat: Added inheritence entity for tag and topic
