+++ src/Entity/Topic.php
@@ -3,81 +3,19 @@
 namespace App\Entity;
 
 use App\Repository\TopicRepository;
-use DateTime;
-use DateTimeImmutable;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
-use Doctrine\DBAL\Types\Types;
 use Doctrine\ORM\Mapping as ORM;
-use Symfony\Component\Uid\Uuid;
-use Symfony\Component\Validator\Constraints as Assert;
 
 #[ORM\Entity(repositoryClass: TopicRepository::class)]
-#[ORM\HasLifecycleCallbacks]
-class Topic
+class Topic extends Classifier
 {
-    #[ORM\Id]
-    #[ORM\Column(type: 'uuid', unique: true)]
-    #[ORM\GeneratedValue(strategy: 'CUSTOM')]
-    #[ORM\CustomIdGenerator(class: 'doctrine.uuid_generator')]
-    private ?Uuid $id = null;
-
-    #[ORM\Column(type: Types::STRING, length: 50, unique: true)]
-    #[Assert\NotBlank(message: "Le nom d'un topic ne peut pas être vide")]
-    #[Assert\Length(max: 50, maxMessage: "Le nom d'un topic ne peut pas dépasser {{ limite }} caractères")]
-    private ?string $name = null;
-
-    #[ORM\Column]
-    private ?\DateTimeImmutable $createdAt = null;
-
-    #[ORM\Column(type: Types::DATETIME_MUTABLE)]
-    private ?\DateTimeInterface $lastUseAt = null;
-
     #[ORM\OneToMany(mappedBy: 'topic', targetEntity: Publication::class)]
     private Collection $publications;
 
     public function __construct()
     {
         $this->publications = new ArrayCollection();
-    }
-
-    public function getId(): ?Uuid
-    {
-        return $this->id;
-    }
-
-    public function getName(): ?string
-    {
-        return $this->name;
-    }
-
-    public function setName(string $name): self
-    {
-        $this->name = $name;
-
-        return $this;
-    }
-
-    public function getCreatedAt(): ?DateTimeImmutable
-    {
-        return $this->createdAt;
-    }
-
-    #[ORM\PrePersist]
-    public function setCreatedAt(): void
-    {
-        $this->createdAt = new DateTimeImmutable();
-    }
-
-    public function getLastUseAt(): ?\DateTimeInterface
-    {
-        return $this->lastUseAt;
-    }
-
-    #[ORM\PrePersist]
-    public function updateLastUseAt(): void
-    {
-        $this->lastUseAt = new DateTime();
     }
 
     /**
